+++ packages/livedoc-mocha/_src/app/reporter/ColorTheme.ts
@@ -27,4 +27,5 @@ export interface ColorTheme {
     dataTableHeader: Chalk;
     dataTable: Chalk;
     docString: Chalk;
+    valuePlaceholders: Chalk
 }