+++ packages/livedoc-mocha/_src/app/reporter/ReporterOptions.ts
@@ -1,8 +1,9 @@
 import { ReporterTheme } from "./ReporterTheme";
 import { ColorTheme } from "./ColorTheme";
+import { DefaultReporter, DefaultColorTheme } from ".";
 
 export class ReporterOptions {
-    public reporter: ReporterTheme;
-    public colors: ColorTheme;
+    public reporter: ReporterTheme = new DefaultReporter();
+    public colors: ColorTheme = new DefaultColorTheme();
     public options: Object;
 }