+++ packages/livedoc-mocha/docs/ReleaseNotes.md
@@ -6,10 +6,12 @@
 
 ### additional with no issue logged
 * ** BREAKING CHANGE** Enforced properly structured Gherkin. Removed some rule violation settings allowing mixing BDD and Gherkin. This needed to be removed to properly support reporting.
+* ** BREAKING CHANGE ** Scenario Outline step titles no longer have their values bound to the example. To get the bound version, the displayTitle can be used. This change was added to support highlighting of values in reporters.
 * Added execution results to model
 * Reorganized and improved Specs and implemented Specs for areas that previously were manual or thew expected exceptions.
 * Added support to run Specs directly and receive an ExecutionResult object with the results of all Specs executed. This feature allowed verification of any scenario including failing ones.
 * Added limited livedoc support to other bdd ui's such as describe. This is to allow stock mocha to make use of future reporting. 
+* Added isolation of global livedoc object for supplying options. It will now be used as the basis for options but other ways to provide options such as via mocha.options and commandline will either extend or override the settings. If possible options should be passed via mocha.options if executing via code otherwise use the command line.
 
 ## 0.3.4
 * Feature: [#52](https://github.com/dotnetprofessional/LiveDoc/issues/52) Support filtering by tags